requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11
django-debug-toolbar==1.8
django-js-asset==1.0.0
django-tinymce==2.7.0
gunicorn==19.4.5
idna==2.6
psycopg2==2.7.3.1
requests==2.18.4
sqlparse==0.2.4
urllib3==1.22
whitenoise==3.3.1
